requirements.txt
# Core Dependencies
google-adk>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
PyYAML>=6.0.0

# Vector Database
chromadb>=0.4.0

# ML/Graph Neural Networks
torch>=2.0.0
torch-geometric>=2.4.0
numpy>=1.24.0
scipy>=1.11.0
networkx>=3.0

# Security Tools
semgrep>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Dashboard UI
fastapi>=0.100.0
uvicorn>=0.22.0
jinja2>=3.1.0

# Utilities
requests>=2.31.0
aiofiles>=23.0.0
